Fleshing out some of the internal flows
Fleshing out some of the internal flows around these concepts, Storage, Security, Contacts and Email

+++ EmailGUI/src/backend/IStoredObjectAPI.java
@@ -9,7 +9,7 @@ public interface IStoredObjectAPI<T> {
 	/*
 	 * The unique key
 	 */
-	int getKey();
+	long getKey();
 	
 	/*
 	 * This helps the storage manager know which table to use
